PDF Summarizer/requirements.txt
streamlit
pypdf
python-dotenv
google-generativeai

